+++ crates/augmented/data/augmented-midi/README.md
@@ -1,10 +1,12 @@
 # augmented-midi
-Implements a MIDI (file) parser using `nom` combinators.
 
-# Specification
+## augmented-midi
+Implements a MIDI (file) parser/serializer using `nom` and `cookie-factory` combinators.
+
+## Specification
 Based on MIDI 1.0 specification. https://www.midi.org/specifications/midi1-specifications/m1-v4-2-1-midi-1-0-detailed-specification-96-1-4.
 
-# License notes
+## License notes
 Bach MIDI file from http://www.piano-midi.de/copy.htm.
 
 Name: Bernd Krueger
@@ -13,3 +15,5 @@ Source: http://www.piano-midi.de
 The distribution or public playback of the files is only allowed under identical license conditions.
 
 The scores are open source.
+
+License: MIT
